feat: apply drag + adjust movement consts
An extra addition in this commit is the additional multiplication of
delta_time in move player. It got tied up in the new const values so it
doesn't make sense to split the commit.

This still seems odd to me, multiplying twice, but I'll leave it in
because it makes the constants easier to manage.