feat!: replace homebrew physics with bevy_xpbd
Use a full physics engine, while retaining the same feel mostly.

Worth noting that xpbd will undergo a algorithm and branding change in
the near future, but they've assured the API will mostly stay the same
so we'll do the upgrade when it happens.

Due to the complexity of this feature, I've chosen to forgo the
slowdown/accel on movement keys up and down respectively. This will be
re-implemented later using the new physics engine.

BREAKING CHANGE: Temporarily remove ability to slow down, or accelerate