feat: add overlay camera for UI outside viewport
Mostly a proof of concept, adding a new camera, set the correct scaling
mode, so that it always stays in the same position, and is scaled to use
pixel art.

In the future this overlay will contain UI elements that are for
platforms that has space left over on the sides.

Currently drawing a square, and placing it outside viewport

TODOs:
- Top position needs figuring out
- Move debug interface somewhere that doesn't interfere with new overlay
  UI